+++ inc/admin-changes.php
@@ -24,12 +24,12 @@ class chg_Posts_to_Articles
         global $wp_post_types;
         $labels = &$wp_post_types['post']->labels;
         $labels->name = 'Blog articles';
-        $labels->singular_name = 'Blog articles';
-        $labels->add_new = 'Add Blog articles';
-        $labels->add_new_item = 'Add Blog articles';
-        $labels->edit_item = 'Edit Blog articles';
+        $labels->singular_name = 'Blog article';
+        $labels->add_new = 'Add Blog article';
+        $labels->add_new_item = 'Add new Blog article';
+        $labels->edit_item = 'Edit Blog article';
         $labels->new_item = 'Blog articles';
-        $labels->view_item = 'View Blog articles';
+        $labels->view_item = 'View Article';
         $labels->search_items = 'Search Blog articles';
         $labels->not_found = 'No Blog articles found';
         $labels->not_found_in_trash = 'No Blog articles found in trash';
@@ -85,31 +85,31 @@ add_action('admin_bar_menu', 'lcm_remove_howdy', 10, 1);
  * Add a link to plugins page under the (Home icon) site name in the admin bar on the frontend.
  * 
  */
-function lcm_add_plugin_link_to_admin_bar() {
-	global $wp_admin_bar;
+// function lcm_add_plugin_link_to_admin_bar() {
+// 	global $wp_admin_bar;
 
-    if ( !is_admin() ) {
-        $wp_admin_bar->add_menu( array(
-            'parent' => 'site-name', // site-name is the ID for home menu so we use it as the parent ID.
-            'id' => 'custom_plugins_link', // You can add any value here as you are adding something very new here.
-            'title' => __( 'Plugins' ), // The anchor text.
-            'href' => home_url() . '/wp-admin/plugins.php', // URL to plugins page.
-            'meta'  => array(
-                'target' => '_self', // Open in another tab.
-            ),
-        ));
-    }
-}
-add_action( 'wp_before_admin_bar_render', 'lcm_add_plugin_link_to_admin_bar' );
+//     if ( !is_admin() ) {
+//         $wp_admin_bar->add_menu( array(
+//             'parent' => 'site-name', // site-name is the ID for home menu so we use it as the parent ID.
+//             'id' => 'custom_plugins_link', // You can add any value here as you are adding something very new here.
+//             'title' => __( 'Plugins' ), // The anchor text.
+//             'href' => home_url() . '/wp-admin/plugins.php', // URL to plugins page.
+//             'meta'  => array(
+//                 'target' => '_self', // Open in another tab.
+//             ),
+//         ));
+//     }
+// }
+// add_action( 'wp_before_admin_bar_render', 'lcm_add_plugin_link_to_admin_bar' );
 
 /**
  * Add a link to Reusable blocks nn Apperance menu.
  * 
  * @since 1.0.1.2
  */
-function add_custom_link_into_appearnace_menu() {
-    global $submenu;
-    $permalink = home_url() . '/wp-admin/edit.php?post_type=wp_block';
-    $submenu['themes.php'][] = array( 'Reusable Blocks', 'manage_categories', $permalink );
-}
-add_action('admin_menu', 'add_custom_link_into_appearnace_menu');
+// function add_custom_link_into_appearnace_menu() {
+//     global $submenu;
+//     $permalink = home_url() . '/wp-admin/edit.php?post_type=wp_block';
+//     $submenu['themes.php'][] = array( 'Reusable Blocks', 'manage_categories', $permalink );
+// }
+// add_action('admin_menu', 'add_custom_link_into_appearnace_menu');