+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) ([Spanish version](https://keepachangelog.com/es-ES/1.0.0/)),and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (or at least I'll try to).
 
+### [Version 1.0.2.3](https://github.com/LuisColome/CoreFunctionalityPlugin/releases/tag/v1.0.2.3) - 2024-15-04
+
+#### Changed
+
+-   New styles to login page.
+
+#### Remove
+
+-   Link to plugins from admin bar (Included from WP 6.5).
+-   Link to Patterns (former reusable blocks). It been included in WP 6.5
+
 ### [Version 1.0.2.2](https://github.com/LuisColome/CoreFunctionalityPlugin/releases/tag/v1.0.2.2) - 2023-15-06
 
 #### Fixed